deleted CardWithFilters Component
